Cleaning up code. Should we use pair and map for pins again?

+++ src/graphics/circuitBoard/components/nodes/node.cpp
@@ -18,12 +18,12 @@ bool Node::isInside(glm::vec2 checkCell) const {
            && checkCell.x <= this->cell.x + float(this->size.x) && checkCell.y <= this->cell.y + float(this->size.y);
 }
 
-uint Node::getInputPinIndex(glm::vec2 absInputPin) {
+uint8_t Node::getInputPinIndex(glm::vec2 absInputPin) {
     const auto iter = std::find(this->inputPins.begin(), this->inputPins.end(), uintVec2(absInputPin - glm::vec2(this->cell)));
     return std::distance(this->inputPins.begin(), iter);
 }
 
-uint Node::getOutputPinIndex(glm::vec2 absOutputPin) {
+uint8_t Node::getOutputPinIndex(glm::vec2 absOutputPin) {
     const auto iter = std::find(this->outputPins.begin(), this->outputPins.end(), uintVec2(absOutputPin - glm::vec2(this->cell)));
     return std::distance(this->outputPins.begin(), iter);
 }
